==> 4_historiographic_analysis.R <==
# load libraries
library(bibliometrix)
library(openxlsx)
library(stringr)
library(plyr)

# load data (optional) - MCESUS serves an input
MAU <- readRDS(file = "data/MCESUS.RDS")
# write.xlsx(MAU, file = "histplot/MCESUS2019.xlsx", row.names=T)

# manually remove remaining duplicated articles
MAU <- subset(MAU, SR != "GIRAUD G, 2016, REV ECON DEV-a")
MAU <- subset(MAU, SR != "IEVDOKYMOV V, 2018, EKON")
MAU <- subset(MAU, SR != "SENETRA A, 2019, ROCZ OCHR SR")
MAU <- subset(MAU, SR != "ZHU Q, 2014, J IND ECOL")

# Tidy up Cited references (top 100 Papers) --> Bibliometrix uses SR to match
MAU$CR <- gsub("[\\:|\\?|\\'|\\’]","",MAU$CR)
MAU$CR <- gsub("[\\-]"," ",MAU$CR)
MAU$CR <- gsub("   |  "," ",MAU$CR)
MAU$CR <- gsub("\\()","",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2015.09.00|THE EXPECTED TRANSITION TO A BALANCED INTERPLAY OF ENVIRONMENTAL AND ECONOMIC SYSTEMS|THE EXPECTED TRANSITION TO A BALANCED INTERPLAY OR ENVIRONMENTAL AND ECONOMIC SYSTEMS",
              ";GHISELLINI P, 2016, J CLEAN PROD;",MAU$CR)

MAU$CR <- gsub("10.1016/J.JCLEPRO.2016.12.048|A NEW SUSTAINABILITY PARADIGM|J. CLEAN. PROD., 143, PP. 757-768",
               ";GEISSDOERFER M, 2017, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("KIRCHHERR J, 2017, RESOUR CONSERV RECY|10.1016/J.RESCONREC.2017.09.005|CONCEPTUALIZING THE CIRCULAR ECONOMY",
              ";KIRCHHERR J, 2017, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("10.1038/NMAT4777|SUSTAINABILITY AND IN SITU MONITORING IN BATTERY DEVELOPMENT",
              ";GREY C, 2016, NAT MATER;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2015.12.042|TOWARDS CIRCULAR ECONOMY IMPLEMENTATION A COMPREHENSIVE REVIEW IN CONTEXT OF MANUFACTURING INDUSTRY",
              ";LIEDER M, 2016, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1007/S10551 015 2693 2|THE CIRCULAR ECONOMY AN INTERDISCIPLINARY EXPLORATION OF THE CONCEPT AND APPLICATION IN A GLOBAL CONTEXT",
              ";MURRAY A, 2017, J BUS ETHICS;",MAU$CR)

MAU$CR <- gsub("10.1016/J.JCLEPRO.2012.11.020|A REVIEW OF THE CIRCULAR ECONOMY IN CHINA MOVING FROM RHETORIC TO IMPLEMENTATION|
              MOVING FROM RHETORIC TO IMPLEMENTATION",
               ";SU B, 2013, J CLEAN PROD;",MAU$CR)

MAU$CR <- gsub("BOCKEN NMP, 2016, J IND PROD ENG","BOCKEN N, 2016, J IND PROD ENG",MAU$CR)
MAU$CR <- sub("10.1080/21681015.2016.1172124|PRODUCT DESIGN AND BUSINESS MODEL STRATEGIES FOR A CIRCULAR ECONOMY",
              ";BOCKEN N, 2016, J IND PROD ENG;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2013.10.051|CURRENT OPTIONS FOR THE VALORIZATION OF FOOD MANUFACTURING WASTE A REVIEW",
              ";MIRABELLA N, 2014, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2011.07.005|TOWARDS A NATIONAL CIRCULAR ECONOMY INDICATOR SYSTEM IN CHINA|TOWARDS A CIRCULAR ECONOMY INDICATOR SYSTEM IN CHINA",
              ";GENG Y, 2012, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.SEPPUR.2016.08.031|RECOVERY AND RECYCLING OF LITHIUM A REVIEW",
              ";SWAIN B, 2017, SEP PURIF TECHNOL;",MAU$CR)

MAU$CR <- sub("10.1039/C6GC02157C|THE E FACTOR 25 YEARS ON THE RISE OF GREEN CHEMISTRY AND SUSTAINABILITY",
              ";SHELDON R, 2017, GREEN CHEM;",MAU$CR)

MAU$CR <- sub("10.1016/J.ECOLECON.2017.06.041|THE CONCEPT AND ITS LIMITATIONS",
              ";KORHONEN J, 2018, ECOL ECON;",MAU$CR)

MAU$CR <- sub("GENG Y, 2008, INT J SUST DEV WORLD|10.3843/SUSDEV.15.36|DEVELOPING THE CIRCULAR ECONOMY IN CHINA CHALLENGES AND OPPORTUNITIES|CIRCULAR ECONOMY IN CHINA CHALLENGES AND OPPORTUNITIES FOR ACHIEVING LEAPFROG DEVELOPMENT",
              ";GENG Y, 2008, INT J SUSTAINABLE DEV WORLD ECOL;",MAU$CR)

MAU$CR <- sub("10.1111/JIEC.12244|HOW CIRCULAR IS THE GLOBAL ECONOMY",
              ";HAAS W, 2015, J IND ECOL;",MAU$CR)

MAU$CR <- sub("10.1016/J.OMEGA.2015.05.015|SUSTAINABLE SUPPLY CHAIN MANAGEMENT AND THE TRANSITION TOWARDS A CIRCULAR ECONOMY",
              ";GENOVESE A, 2017, OMEGA;",MAU$CR)

MAU$CR <- sub("10.1016/J.BIORTECH.2016.03.130|WASTE BIOREFINERY MODELS TOWARDS SUSTAINABLE CIRCULAR BIOECONOMY CRITICAL REVIEW AND FUTURE PERSPECTIVES",
              ";VENKATA M S, 2016, BIORESOUR TECHNOL;",MAU$CR)

MAU$CR <- sub("10.3390/SU8010043|DESIGNING THE BUSINESS MODELS FOR CIRCULAR ECONOMY",
              ";LEWANDOWSKI M, 2016, SUSTAINABILITY;",MAU$CR)

MAU$CR <- sub("10.1016/J.SAA.2012.05.083|AROMAL SA, 2012, SPECTROC ACTA PT A-MOLEC BIOMOLEC SPECTR|GREEN SYNTHESIS OF GOLD",
              ";AROMAL S, 2012, SPECTROC ACTA PT A-MOLEC BIOMOLEC SPECTR;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2014.01.028|PRODUCTS THAT GO ROUND EXPLORING PRODUCT LIFE EXTENSION THROUGH DESIGN",
              ";BAKKER C, 2014, J CLEAN PROD;",MAU$CR)

MAU$CR <- gsub("MATHEWS JA, 2011, J IND ECOL","MATHEWS J, 2011, J IND ECOL",MAU$CR)
MAU$CR <- sub("10.1111/J.1530 9290.2011.00332.X|PROGRESS TOWARD A CIRCULAR ECONOMY IN CHINA|PROGRESS TOWARDS A CIRCULAR ECONOMY THE DRIVERS AND INHIBITORS OF ECO INDUSTRIAL INITIATIVE|PROGRESS TOWARDS A CIRCULAR ECONOMY IN CHINA THE DRIVERS",
              ";MATHEWS J, 2011, J IND ECOL;",MAU$CR)

MAU$CR <- sub("SAUVE S, 2016, ENVIRON DEV|10.1016/J.ENVDEV.2015.09.002|ENVIRONMENTAL SCIENCES, SUSTAINABLE DEVELOPMENT AND CIRCULAR ECONOMY|ENVIRONMENTAL SCIENCES SUSTAINABLE DEVELOPMENT AND CIRCULAR ECONOMY ALTERNATIVE CONCEPTS FOR TRANS DISCIPLINARY RESEARCH",
              ";SAUV S, 2016, ENVIRON DEV;",MAU$CR)

MAU$CR <- gsub("ANDERSEN M.S.|ANDERSEN MS|ANDERSEN M. S.|ANDERSEN, M.S.","ANDERSEN M",MAU$CR)
MAU$CR <- sub("ANDERSEN M, 2007, SUSTAINABILITY SCI|10.1007/S11625 006 0013 6|AN INTRODUCTORY NOTE ON THE ENVIRONMENTAL ECONOMICS OF THE CIRCULAR ECONOMY",
              ";ANDERSEN M, 2007, SUSTAIN SCI;",MAU$CR)

MAU$CR <- sub("HUANG SL","HUANG S",MAU$CR)

MAU$CR <- sub("WITJES S, 2016, RESOUR CONSERV RECY|10.1016/J.RESCONREC.2016.04.015|PROPOSING A FRAMEWORK LINKING SUSTAINABLE PUBLIC PROCUREMENT AND SUSTAINABLE BUSINESS MODELS",
              ";WITJES S, 2016, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2010.06.001|CREATING INTEGRATED BUSINESS AND ENVIRONMENTAL VALUE WITHIN THE CONTEXT OF CHINA",
              ";PARK J, 2010, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1080/03085147.2015.1013353|INTERROGATING THE CIRCULAR ECONOMY",
              ";GREGSON N, 2015, ECON SOC;",MAU$CR)

MAU$CR <- sub("10.1016/J.ENVRES.2017.03.010|SEWAGE SLUDGE DISPOSAL STRATEGIES FOR SUSTAINABLE DEVELOPMENT",
              ";KACPRZAK M, 2017, ENVIRON RES;",MAU$CR)

MAU$CR <- sub("KALMYKOVA Y, 2018, RESOUR CONSERV RECY|10.1016/J.RESCONREC.2017.10.034|FROM REVIEW OF THEORIES AND PRACTICES TO DEVELOPMENT OF IMPLEMENTATION TOOLS",
              ";KALMYKOVA Y, 2018, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2010.09.018|ECOLOGICAL UTILIZATION OF LEATHER TANNERY WASTE WITH CIRCULAR ECONOMY MODEL",
              ";HU J, 2011, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2011.01.003|THE DYNAMICS OF INDUSTRIAL SYMBIOSIS A PROPOSAL FOR A CONCEPTUAL FRAMEWORK BASED UPON A COMPREHENSIVE LITERATURE REVIEW",
              ";BOONS F, 2011, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("ZHIJUN F., 2007, SUSTAIN SCI|ZHIJUN F, 2007, SUSTAIN SCI|10.1007/S11625 006 0018 1|PUTTING A CIRCULAR ECONOMY INTO PRACTICE IN CHINA",
              ";ZHIJUN F, 2007, SUSTAINABILITY SCI;",MAU$CR)

MAU$CR <- sub("10.1073/PNAS.1613773114|GLOBAL SOCIOECONOMIC MATERIAL STOCKS RISE 23 FOLD OVER THE 20TH CENTURY AND REQUIRE HALF OF ANNUAL RESOURCE USE",
              ";KRAUSMANN F, 2017, PROC NATL ACAD SCI U S A;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2015.02.051|THE POSSIBLE USE OF SEWAGE SLUDGE ASH",
              ";SMOL M, 2015, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("LINDER M., 2015, BUS STRATEG ENVIRON|10.1002/BSE.1906|CIRCULAR BUSINESS MODEL INNOVATION INHERENT UNCERTAINTIES",
              ";LINDER M, 2017, BUS STRATEG ENVIRON;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2016.12.055|CIRCULAR ECONOMY FOR THE BUILT ENVIRONMENT A RESEARCH FRAMEWORK",
              ";POMPONI F, 2017, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2015.07.133|SHARING CITIES AND SUSTAINABLE CONSUMPTION AND PRODUCTION TOWARDS AN INTEGRATED FRAMEWORK",
              ";COHEN B, 2016, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2017.12.111|CIRCULAR ECONOMY AS AN ESSENTIALLY CONTESTED CONCEPT",
              ";KORHONEN J, 2018, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1021/ES201904C|MOVING TOWARD THE CIRCULAR ECONOMY THE ROLE OF STOCKS IN THE CHINESE STEEL CYCLE",
              ";PAULIUK S, 2012, ENVIRON SCI TECHNOL;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2017.09.053|GREEN, CIRCULAR, BIO ECONOMY A COMPARATIVE ANALYSIS OF SUSTAINABILITY AVENUES",
              ";D A D, 2017, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("SHELDON RA, 2018, ACS SUSTAINABLE CHEM ENG|10.1021/ACSSUSCHEMENG.7B03505|METRICS OF GREEN CHEMISTRY AND SUSTAINABILITY",
              ";SHELDON R, 2018, ACS SUSTAINABLE CHEM ENG;",MAU$CR)

MAU$CR <- sub("WU HQ, 2014, RESOUR CONSERV RECY|10.1016/J.RESCONREC.2013.10.003|EFFECTIVENESS OF THE POLICY OF CIRCULAR ECONOMY IN CHINA",
              ";WU H, 2014, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("DODSON JR, 2015, GREEN CHEM|10.1039/C4GC02483D|BIODERIVED MATERIALS AS A GREEN ROUTE FOR PRECIOUS",
              ";DODSON J, 2015, GREEN CHEM;",MAU$CR)

MAU$CR <- sub("10.1016/10.1016/J.JCLEPRO.2016.10.196|MEASURING CIRCULAR ECONOMY STRATEGIES THROUGH INDEX METHODS A CRITICAL ANALYSIS",
              ";ELIA V, 2017, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2017.12.112|HOW DO SCHOLARS APPROACH THE CIRCULAR ECONOMY",
              ";MERLI R, 2018, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.SCITOTENV.2010.07.081|EMERGY ANALYSIS OF AN INDUSTRIAL PARK",
              ";GENG Y, 2010, SCI TOTAL ENVIRON;",MAU$CR)

MAU$CR <- sub("PRIETO SANDOVAL V, 2018, J CLEAN PROD|10.1016/J.JCLEPRO.2017.12.224|TOWARDS A CONSENSUS ON THE CIRCULAR ECONOMY",
              ";PRIETO S V, 2018, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2015.12.020|RESOURCE RECOVERY FROM POST CONSUMER WASTE IMPORTANT LESSONS FOR THE UPCOMING CIRCULAR ECONOMY",
              ";SINGH J, 2016, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2017.09.047|TOWARDS A NEW TAXONOMY OF CIRCULAR ECONOMY BUSINESS MODELS",
              ";URBINATI A, 2017, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.3390/SU8090937|A CONCEPTUAL FRAMEWORK FOR CIRCULAR DESIGN",
              ";MORENO M, 2016, SUSTAINABILITY;",MAU$CR)

MAU$CR <- sub("EGLE L, 2015, RESOUR CONSERV RECY|10.1016/J.RESCONREC.2015.09.016|OVERVIEW AND DESCRIPTION OF TECHNOLOGIES FOR RECOVERING PHOSPHORUS",
              ";EGLE L, 2015, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("SCHEEPENS AE, 2016, J CLEAN PROD|10.1016/J.JCLEPRO.2015.05.075|MAKING WATER TOURISM MORE SUSTAINABLE",
              ";SCHEEPENS A, 2016, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("WINKLER H, 2011, CIRP J MANUF SCI TEC|10.1016/J.CIRPJ.2011.05.001|CLOSED LOOP PRODUCTION SYSTEMS A SUSTAINABLE SUPPLY CHAIN APPROACH|
               CLOSEDLOOP PRODUCTION SYSTEMSA SUSTAINABLE SUPPLY CHAIN APPROACH",
              ";WINKLER H, 2011, CIRP J MANUF SCI TECHNOL;",MAU$CR)

MAU$CR <- sub("10.1016/10.1098/RSTA.2014.0177|CO2 UTILIZATION AN ENABLING ELEMENT TO",
              ";AMPELLI C, 2015, PHILOS TRANS R SOC A MATH PHYS ENG SCI;",MAU$CR)

MAU$CR <- sub("FERNANDEZ JE, 2007, J IND ECOL|10.1162/JIE.2007.1199|RESOURCE CONSUMPTION OF NEW URBAN CONSTRUCTION IN CHINA",
              ";FERNNDEZ J, 2007, J IND ECOL;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2016.08.024|GREEN ECONOMY AND RELATED CONCEPTS",
              ";LOISEAU E, 2016, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.3389/FMICB.2016.02106|BY BIOLOGICAL TECHNOLOGIES",
              ";PUYOL D, 2017, FRONT MICROBIOL;",MAU$CR)

MAU$CR <- sub("10.1007/S13280 014 0614 8|WITHERS PJA, 2015, AMBIO|STEWARDSHIP TO TACKLE GLOBAL PHOSPHORUS INEFFICIENCY",
              ";WITHERS P, 2015, AMBIO;",MAU$CR)

MAU$CR <- sub("10.1039/C7GC01496A|CHEMICALLY RECYCLABLE POLYMERS A CIRCULAR ECONOMY APPROACH TO SUSTAINABILITY",
              ";HONG M, 2017, GREEN CHEM;",MAU$CR)

MAU$CR <- sub("REIKE D, 2018, RESOUR CONSERV RECY|10.1016/J.RESCONREC.2017.08.027|NEW OR REFURBISHED AS CE",
              ";REIKE D, 2018, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("XUE B, 2010, RESOUR CONSERV RECY|10.1016/J.RESCONREC.2010.05.010|SURVEY OF OFFICIALS AWARENESS ON CIRCULAR ECONOMY DEVELOPMENT IN CHINA, BASED ON MUNICIPAL AND COUNTY LEVEL|SURVEY OF OFFICIALS AWARENESS ON CIRCULAR ECONOMY DEVELOPMENT IN CHINA BASED ON MUNICIPAL AND COUNTY LEVEL",
              ";XUE B, 2010, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("DE LOS RIOS IC, 2017, J CLEAN PROD|10.1016/J.JCLEPRO.2016.10.130|SKILLS AND CAPABILITIES FOR A SUSTAINABLE AND CIRCULAR ECONOMY",
              ";DE L R I, 2017, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.3390/CATAL6120196|NEW FRONTIERS IN THE CATALYTIC SYNTHESIS OF LEVULINIC ACID",
              ";ANTONETTI C, 2016, CATALYSTS;",MAU$CR)

MAU$CR <- sub("EVIDENCE FROM THE EUROPEAN UNION|10.1016/J.ECOLECON.2018.04.028",
              ";KIRCHHERR J, 2018, ECOL ECON;",MAU$CR)

MAU$CR <- sub("JABBOUR ABLD, 2018, ANN OPER RES|10.1007/S10479 018 2772 8|A PROPOSED RESEARCH AGENDA AND ORIGINAL ROADMAP FOR SUSTAINABLE OPERATIONS",
              ";LOPES D A, 2018, ANN OPER RES;",MAU$CR)

MAU$CR <- sub("PARK JY, 2014, J ENVIRON MANAGE|10.1016/J.JENVMAN.2013.11.053|ESTABLISHING AND TESTING THE REUSE POTENTIAL INDICATOR FOR MANAGING WASTES AS RESOURCES",
              ";PARK J, 2014, J ENVIRON MANAGE;",MAU$CR)

MAU$CR <- sub("10.1016/J.TECHFORE.2016.09.021|UNLOCKING VALUE FOR A CIRCULAR ECONOMY THROUGH 3D PRINTING",
              ";DESPEISSE M, 2017, TECHNOL FORECAST SOC CHANGE;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2006.07.019|AN ECONOMIC INSTRUMENT FOR ZERO WASTE, ECONOMIC GROWTH AND SUSTAINABILITY",
              ";GREYSON J, 2007, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("HOBSON K, 2016, PROG HUM GEOG|10.1177/0309132514566342|LOCATING GENERATIVE SPACES FOR THE CIRCULAR ECONOMY",
              ";HOBSON K, 2016, PROG HUM GEOGR;",MAU$CR)

MAU$CR <- sub("SHELDON RA, 2016, J MOL CATAL A CHEM|10.1016/J.MOLCATA.2016.01.013|GREEN CHEMISTRY, CATALYSIS AND VALORIZATION OF WASTE BIOMASS",
              ";SHELDON R, 2016, J MOL CATAL A CHEM;",MAU$CR)

MAU$CR <- sub("PAULIUK S, 2018, RESOUR CONSERV RECY|10.1016/J.RESCONREC.2017.10.019|CRITICAL APPRAISAL OF THE CIRCULAR ECONOMY STANDARD BS 8001",
              ";PAULIUK S, 2018, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("YANG SL, 2008, RESOUR CONSERV RECY|10.1016/J.RESCONREC.2007.11.008|A CASE STUDY OF INDUSTRIAL SYMBIOSIS NANNING SUGAR CO",
              ";YANG S, 2008, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("SUSTAINABLE STRATEGY FOR CIRCULAR BIOECONOMY|10.1016/J.BIORTECH.2017.07.176",
              "DAHIYA S, 2018, BIORESOUR TECHNOL",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2018.04.159|BUSINESS MODELS AND SUPPLY CHAINS FOR THE CIRCULAR ECONOMY",
              "GEISSDOERFER M, 2018, J CLEAN PROD",MAU$CR)

MAU$CR <- sub("10.1080/00207543.2017.1402141|A SYSTEMATIC REVIEW ON DRIVERS BARRIERS AND PRACTICES TOWARDS CIRCULAR ECONOMY",
              ";GOVINDAN K, 2018, INT J PROD RES;",MAU$CR)

MAU$CR <- sub("LIU Y, 2014, RESOUR CONSERV RECY|10.1016/J.RESCONREC.2014.04.002|AN EXPLORATION OF FIRMS AWARENESS AND BEHAVIOR OF DEVELOPING CIRCULAR ECONOMY AN EMPIRICAL RESEARCH IN CHINA",
              ";LIU Y, 2014, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("10.1016/J.ECOLECON.2017.08.001|DE JESUS A, 2018, ECOL ECON|LOST IN TRANSITION",
              ";DE J A, 2018, ECOL ECON;",MAU$CR)

MAU$CR <- sub("10.1016/J.FUTURES.2016.05.012|RADICAL SOCIAL TRANSFORMATION IN A RESOURCE SCARCE",
              ";HOBSON K, 2016, FUTURES;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2017.10.003|DO CIRCULAR ECONOMY BUSINESS MODELS CAPTURE INTENDED ENVIRONMENTAL",
              ";MANNINEN K, 2018, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("NASIR MHA, 2017, INT J PROD ECON|10.1016/J.IJPE.2016.06.008|COMPARING LINEAR AND CIRCULAR SUPPLY CHAINS A CASE STUDY FROM THE CONSTRUCTION INDUSTRY",
              ";NASIR M, 2017, INT J PROD ECON;",MAU$CR)

MAU$CR <- sub("10.3390/SU8010069|TRANSITION TOWARDS CIRCULAR ECONOMY IN THE FOOD SYSTEM",
              ";JURGILEVICH A, 2016, SUSTAINABILITY;",MAU$CR)

MAU$CR <- sub("ZHU JB, 2018, SCIENCE|10.1126/SCIENCE.AAR5498|A SYNTHETIC POLYMER SYSTEM WITH REPEATABLE CHEMICAL RECYCLABILITY",
              ";ZHU J, 2018, SCIENCE;",MAU$CR)

MAU$CR <- sub("LIGUORI R, 2016, BIORESOURCE TECHNOL|10.1016/J.BIORTECH.2016.04.054|BIOLOGICAL PROCESSES FOR ADVANCING LIGNOCELLULOSIC WASTE BIOREFINERY BY ADVOCATING CIRCULAR ECONOMY",
              ";LIGUORI R, 2016, BIORESOUR TECHNOL;",MAU$CR)

MAU$CR <- sub("VAN WEELDEN E, 2016, J CLEAN PROD|10.1016/J.JCLEPRO.2015.11.065|PAVING THE WAY TOWARDS CIRCULAR CONSUMPTION EXPLORING CONSUMER ACCEPTANCE",
              ";VAN W E, 2016, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("JIAO WT, 2014, J CLEAN PROD|10.1016/J.JCLEPRO.2013.12.050|TOWARD A RESEARCH AGENDA FOR POLICY INTERVENTION AND FACILITATION TO ENHANCE INDUSTRIAL SYMBIOSIS BASED ON A COMPREHENSIVE LITERATURE REVIEW",
              ";JIAO W, 2014, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("VAN BUREN N, 2016, SUSTAINABILITY|10.3390/SU8070647|TOWARDS A CIRCULAR ECONOMY THE ROLE OF DUTCH LOGISTICS INDUSTRIES AND GOVERNMENTS",
              ";VAN B N, 2016, SUSTAINABILITY;",MAU$CR)

MAU$CR <- sub("10.1111/J.1530 9290.2011.00443.X|METHODOLOGICAL ASPECTS OF APPLYING LIFE CYCLE ASSESSMENT TO INDUSTRIAL SYMBIOSES",
              ";MATTILA T, 2012, J IND ECOL;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2017.11.064|HOMRICH AS, 2018, J CLEAN PROD|THE CIRCULAR ECONOMY UMBRELLA TRENDS AND GAPS ON INTEGRATING PATHWAYS",
              ";HOMRICH A, 2018, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("MA SH, 2014, J CLEAN PROD|10.1016/J.JCLEPRO.2013.10.008|MODE OF CIRCULAR ECONOMY IN CHINAS IRON AND STEEL INDUSTRY",
              ";MA S, 2014, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2013.09.002|THE CASE OF TIANJIN",
              ";YU C, 2014, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("ZENG HX, 2017, J CLEAN PROD|10.1016/J.JCLEPRO.2016.10.093|INSTITUTIONAL PRESSURES, SUSTAINABLE SUPPLY CHAIN MANAGEMENT, AND CIRCULAR ECONOMY CAPABILITY",
              ";ZENG H, 2017, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.TIBTECH.2015.12.009|THE CLOSURE OF THE CYCLE ENZYMATIC SYNTHESIS",
              ";PELLIS A, 2016, TRENDS BIOTECHNOL;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2017.07.100|METRICS FOR OPTIMIZING THE MULTI ",
              ";IACOVIDOU E, 2017, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("FRENKEN K, 2017, PHILOS T R SOC A|10.1098/RSTA.2016.0367|POLITICAL ECONOMIES AND ENVIRONMENTAL FUTURES FOR THE SHARING ECONOMY",
              ";FRENKEN K, 2017, PHILOS TRANS R SOC A MATH PHYS ENG SCI;",MAU$CR)

MAU$CR <- sub("10.1016/0378 11279503574 T|BARROS AC, 1995, FOR ECOL MANAGE|LOGGING ALONG THE AMAZON RIVER AND ESTUARY PATTERNS",
              ";BARROS A, 1995, FOR ECOL MANAGE;",MAU$CR)

MAU$CR <- sub("GEORGE DAR, 2015, ENVIRON MODELL SOFTW|10.1016/J.ENVSOFT.2015.06.014|A CIRCULAR ECONOMY MODEL OF ECONOMIC GROWTH",
              ";GEORGE D, 2015, ENVIRON MODEL SOFTW;",MAU$CR)

MAU$CR <- sub("10.1007/S10668 016 9760 4|THE CHALLENGES OF WATER, WASTE AND CLIMATE CHANGE IN CITIES",
              ";KOOP S, 2017, ENVIRON DEV SUSTAINABILITY;",MAU$CR)

MAU$CR <- sub("NESS D, 2008, INT J SUST DEV WORLD|10.3843/SUSDEV.15.42A|SUSTAINABLE URBAN INFRASTRUCTURE IN CHINA TOWARDS A FACTOR",
              ";NESS D, 2008, INT J SUSTAINABLE DEV WORLD ECOL;",MAU$CR)

MAU$CR <- sub("HUYSMAN S, 2017, RESOUR CONSERV RECY|10.1016/J.RESCONREC.2017.01.013|PERFORMANCE INDICATORS FOR A CIRCULAR ECONOMY A CASE STUDY ON POST",
              ";HUYSMAN S, 2017, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("10.3390/SU9091602|SUPPLY CHAIN CONFIGURATIONS IN THE CIRCULAR ECONOMY A SYSTEMATIC LITERATURE REVIEW",
              ";MASI D, 2017, SUSTAINABILITY;",MAU$CR)

MAU$CR <- sub("MATUS KJM, 2012, J CLEAN PROD|10.1016/J.JCLEPRO.2012.03.033|GREEN CHEMISTRY AND GREEN ENGINEERING IN CHINA DRIVERS, POLICIES AND BARRIERS TO INNOVATION",
              ";MATUS K, 2012, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2017.11.063|MOKTADIR MA, 2018, J CLEAN PROD|A PERSPECTIVE OF LEATHER INDUSTRIES IN BANGLADESH",
              ";MOKTADIR M, 2018, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("NOBRE GC, 2017, SCIENTOMETRICS|10.1007/S11192 017 2281 6|SCIENTIFIC LITERATURE ANALYSIS ON BIG DATA AND INTERNET OF THINGS APPLICATIONS ON CIRCULAR ECONOMY A BIBLIOMETRIC STUDY",
              ";NOBRE G, 2017, SCIENTOMETRICS;",MAU$CR)

MAU$CR <- sub("STAHEL WR, 2013, PHILOS T R SOC A|10.1098/RSTA.2011.0567|POLICY FOR MATERIAL EFFICIENCY",
              ";STAHEL W, 2013, PHILOS TRANS R SOC A MATH PHYS ENG SCI;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2017.09.002|A PATHWAY TO CIRCULAR ECONOMY DEVELOPING A CONCEPTUAL FRAMEWORK FOR COMPLEX VALUE ASSESSMENT OF RESOURCES RECOVERED FROM WASTE",
              ";IACOVIDOU E, 2017, J CLEAN PROD-a;",MAU$CR)

MAU$CR <- sub("HAZEN BT, 2017, BUS STRATEG ENVIRON|HAZEN B, 2017, BUS STRATEG ENVIRON|10.1002/BSE.1929|REMANUFACTURING FOR THE CIRCULAR ECONOMY AN EXAMINATION OF CONSUMER SWITCHING BEHAVIOR",
              ";HAZEN B, 2017, BUS STRATEGY ENVIRON;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2017.06.093|SUSTAINABLE SUPPLY CHAIN MANAGEMENT PRACTICES, SUPPLY CHAIN DYNAMIC CAPABILITIES, AND ENTERPRISE PERFORMANCE",
              ";HONG J, 2018, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.BIORTECH.2016.01.091|NEW PERSPECTIVES FOR THE DESIGN OF SUSTAINABLE BIOPROCESSES FOR PHOSPHORUS RECOVERY FROM WASTE",
              ";TARAYRE C, 2016, BIORESOUR TECHNOL;",MAU$CR)

MAU$CR <- sub("ANDREWS D., 2015, LOCAL EC|10.1177/0269094215578226|THE CIRCULAR ECONOMY, DESIGN THINKING AND EDUCATION FOR SUSTAINABILITY",
              ";ANDREWS D, 2015, LOCAL ECON;",MAU$CR)

MAU$CR <- sub("ANTIKAINEN M, 2016, TECHNOL INNOV MANAG|A FRAMEWORK FOR SUSTAINABLE CIRCULAR BUSINESS MODEL INNOVATION",
              ";ANTIKAINEN M, 2016, TECHNOL INNOV MANAG REV;",MAU$CR)

MAU$CR <- sub("10.1002/IEAM.1614|BEYOND THE THROWAWAY SOCIETY A LIFE CYCLE BASED ASSESSMENT OF THE ENVIRONMENTAL BENEFIT OF REUSE",
              ";CASTELLANI V, 2015, INTEGR ENVIRON ASSESS MANAGE;",MAU$CR)

MAU$CR <- sub("10.1016/J.PARTIC.2012.11.001|PROCESS ENGINEERING IN CIRCULAR ECONOMY",
              ";REH L, 2013, PARTICUOLOGY",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2016.11.009|DEVELOPING SUSTAINABLE BUSINESS EXPERIMENTATION CAPABILITY",
              ";WEISSBROD I, 2017, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.RESCONREC.2017.07.031|ISILDAR A, 2018, RESOUR CONSERV RECY|ELECTRONIC WASTE AS A SECONDARY SOURCE OF CRITICAL METALS",
              ";ILDAR A, 2018, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("10.21894/JOPR.2018.0030|OIL PALM ECONOMIC PERFORMANCE IN MALAYSIA",
              ";KUSHAIRI A, 2018, J OIL PALM RES",MAU$CR)

MAU$CR <- sub("10.1186/S13068-017-0761-9|MULTIPRODUCT BIOREFINERIES FROM LIGNOCELLULOSES",
              ";FARZAD S, 2017, BIOTECHNOL BIOFUELS",MAU$CR)

MAU$CR <- sub("FRANKLIN JOHNSON E, 2016, J CLEAN PROD|10.1016/J.JCLEPRO.2016.05.023|RESOURCE DURATION AS A MANAGERIAL INDICATOR FOR CIRCULAR ECONOMY PERFORMANCE|
              FRANKLIN-JOHNSON E, 2016, J CLEAN PROD",
              ";FRANKLIN J E, 2016, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("MA SJ, 2015, J CLEAN PROD|10.1016/J.JCLEPRO.2014.10.059|A CASE STUDY OF A PHOSPHORUS CHEMICAL FIRM",
              ";MA S, 2015, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("NIERO M, 2016, RESOUR CONSERV RECY|10.1016/J.RESCONREC.2016.06.023|CIRCULAR ECONOMY TO BE OR NOT TO BE IN A CLOSED PRODUCT LOOP",
              ";NIERO M, 2016, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2019.01.036|PIERONI MPP, 2019, J CLEAN PROD|BUSINESS MODEL INNOVATION FOR CIRCULAR ECONOMY AND SUSTAINABILITY",
              ";PIERONI M, 2019, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2015.09.022|SUSTAINABILITY IN THE EU CEMENT INDUSTRY THE ITALIAN AND GERMAN",
              ";SUPINO S, 2016, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("BAI CG, 2015, IND MANAGE DATA SYST|10.1108/IMDS 09 2014 0258|CORPORATE SUSTAINABILITY DEVELOPMENT IN CHINA REVIEW AND ANALYSIS",
              ";BAI C, 2015, IND MANAGE DATA SYST;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2016.12.038|INSTITUTIONAL INCENTIVES IN CIRCULAR ECONOMY TRANSITION THE CASE OF MATERIAL USE IN THE DUTCH TEXTILE INDUSTRY",
              ";FISCHER A, 2017, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2016.11.041|EVALUATING THE COMPREHENSIVE BENEFIT OF ECO INDUSTRIAL PARKS BY EMPLOYING MULTICRITERIA DECISION MAKING APPROACH FOR CIRCULAR ECONOMY|EVALUATING THE COMPREHENSIVE BENEFIT OF ECO INDUSTRIAL PARKS BY EMPLOYING MULTI CRITERIA DECISION MAKING APPROACH FOR CIRCULAR ECONOMY",
              ";ZHAO H, 2017, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.WASMAN.2017.10.033|PLASTIC PACKAGING WASTE IN FINLAND",
              ";DAHLBO H, 2018, WASTE MANAGE;",MAU$CR)

MAU$CR <- sub("SUN L, 2017, RESOUR CONSERV RECY|10.1016/J.SPC.2015.06.004|ECO BENEFITS ASSESSMENT ON URBAN INDUSTRIAL SYMBIOSIS BASED ON MATERIAL FLOWS ANALYSIS AND EMERGY EVALUATION APPROACH|
              ECOBENEFITS ASSESSMENT ON URBAN INDUSTRIAL SYMBIOSIS BASED ON MATERIAL FLOWS ANALYSIS AND EMERGY EVALUATION APPROACH",
              ";SUN L, 2017, RESOUR CONSERV RECYCL;",MAU$CR)

MAU$CR <- sub("10.1016/J.JCLEPRO.2017.09.056|FRANCO MA, 2017, J CLEAN PROD|A DYNAMIC VIEW OF INCUMBENTS",
              ";FRANCO M, 2017, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("10.1016/J.APENERGY.2016.02.084|TOWARD A FUNCTIONAL INTEGRATION OF ANAEROBIC DIGESTION AND PYROLYSIS FOR A SUSTAINABLE RESOURCE MANAGEMENT",
              ";MONLAU F, 2016, APPL ENERGY;",MAU$CR)

MAU$CR <- sub("10.1016/J.CONBUILDMAT.2016.03.144",
              ";OZER H, 2016, CONSTR BUILD MATER;",MAU$CR)

MAU$CR <- sub("SAAVEDRA YMB, 2018, J CLEAN PROD|10.1016/J.JCLEPRO.2017.09.260|THEORETICAL CONTRIBUTION OF INDUSTRIAL ECOLOGY TO CIRCULAR ECONOMY",
              ";SAAVEDRA Y, 2018, J CLEAN PROD;",MAU$CR)

MAU$CR <- sub("SHI PJ, 2009, J NAT DISASTERS|THEORY AND PRACTICE ON DISASTER SYSTEM RESEARCH",
              ";SHI P, 2009, J NAT DISASTERS",MAU$CR)

MAU$CR <- sub("WALLS JL, 2015, ORGAN ENVIRON|10.1177/1086026615575333|ORGANIZATIONAL PERSPECTIVES OF INDUSTRIAL SYMBIOSIS A REVIEW AND SYNTHESIS",
              ";WALLS J, 2015, ORGAN ENVIRON;",MAU$CR)

MAU$CR <- sub("TUKKER A., 2013, J CLEAN PROD|10.1016/J.JCLEPRO.2013.11.049|PRODUCT SERVICES FOR A RESOURCE EFFICIENT AND CIRCULAR ECONOMY|PRODUCT SERVICES FOR A RESOURCE EF FI CIENT AND CIRCULAR ECONOMY E",
              ";TUKKER A, 2015, J CLEAN PROD;",MAU$CR)

# wrote excel file
write.xlsx(MAU, file = "histplot/MAU.xlsx",row.names=T)

local_citations<-localCitations(MAU, T)
# write.xlsx(local_citations[["Papers"]], file = "histplot/local_citations_papers.xlsx", row.names=T)
# write.xlsx(local_citations[["M"]], file = "histplot/local_citations_M.xlsx", row.names=T)
local_citations30 <- subset(local_citations[["M"]], subset = LCS > 40)

#Historical Analysis -----------------------------------------------------------------------------------------------
hist_result <- histNetwork(MAU, min.citations = 10, sep = ". ")

# check intra-network citations
INGhisellini <- count(grep("GHISELLINI P, 2016, J CLEAN PROD",local_citations30[["CR"]]))
sum(INGhisellini$freq)

INGeissdoerfer <- rbind(count(grep("GEISSDOERFER M, 2017, J CLEAN PROD",local_citations30[["CR"]])))
sum(INGeissdoerfer$freq)

INLieder <- count(grep("LIEDER M, 2016, J CLEAN PROD",local_citations30[["CR"]]))
sum(INLieder$freq)

INKirchherr <- count(grep("KIRCHHERR J, 2017, RESOUR CONSERV RECYCL",local_citations30[["CR"]]))
sum(INKirchherr$freq)

INMurray <- count(grep("MURRAY A, 2017, J BUS ETHICS",local_citations30[["CR"]]))
sum(INMurray$freq)

INSu <- count(grep("SU B, 2013, J CLEAN PROD",local_citations30[["CR"]]))
sum(INSu$freq)

INBocken <- count(grep("BOCKEN N, 2016, J IND PROD ENG",local_citations30[["CR"]]))
sum(INBocken$freq)

INGeng <- count(grep("GENG Y, 2012, J CLEAN PROD",local_citations30[["CR"]]))
sum(INGeng$freq)

INKorhonen <- count(grep("KORHONEN J, 2018, ECOL ECON",local_citations30[["CR"]]))
sum(INKorhonen$freq)

INGeng2 <- count(grep("GENG Y, 2008, INT J SUSTAINABLE DEV WORLD ECOL",local_citations30[["CR"]]))
sum(INGeng2$freq)

INAndersen <- count(grep("ANDERSEN M, 2007, SUSTAIN SCI",local_citations30[["CR"]]))
sum(INAndersen$freq)

INLewandowski <- count(grep("LEWANDOWSKI M, 2016, SUSTAINABILITY",local_citations30[["CR"]]))
sum(INLewandowski$freq)

INGenovese <- count(grep("GENOVESE A, 2017, OMEGA",local_citations30[["CR"]]))
sum(INGenovese$freq)

INMathews <- count(grep("MATHEWS J, 2011, J IND ECOL",local_citations30[["CR"]]))
sum(INMathews$freq)

INHaas <- count(grep("HAAS W, 2015, J IND ECOL",local_citations30[["CR"]]))
sum(INHaas$freq)

INSauv <- count(grep("SAUV S, 2016, ENVIRON DEV",local_citations30[["CR"]]))
sum(INSauv$freq)

INGregson <- count(grep("GREGSON N, 2015, ECON SOC",local_citations30[["CR"]]))
sum(INGregson$freq)

INBakker <- count(grep("BAKKER C, 2014, J CLEAN PROD",local_citations30[["CR"]]))
sum(INBakker$freq)

INLinder <- count(grep("LINDER M, 2017, BUS STRATEG ENVIRON",local_citations30[["CR"]]))
sum(INLinder$freq)

INKalmykova <- count(grep("KALMYKOVA Y, 2018, RESOUR CONSERV RECYCL",local_citations30[["CR"]]))
sum(INKalmykova$freq)

INZhijun <- count(grep("ZHIJUN F, 2007, SUSTAINABILITY SCI",local_citations30[["CR"]]))
sum(INZhijun$freq)

INWitjes <- count(grep("WITJES S, 2016, RESOUR CONSERV RECYCL",local_citations30[["CR"]]))
sum(INWitjes$freq)

INPomponi <- count(grep("POMPONI F, 2017, J CLEAN PROD",local_citations30[["CR"]]))
sum(INPomponi$freq)

INPark <- count(grep("PARK J, 2010, J CLEAN PROD",local_citations30[["CR"]]))
sum(INPark$freq)

INUrbinati <- count(grep("URBINATI A, 2017, J CLEAN PROD",local_citations30[["CR"]]))
sum(INUrbinati$freq)

INKorhonen2 <- count(grep("KORHONEN J, 2018, J CLEAN PROD",local_citations30[["CR"]]))
sum(INKorhonen2$freq)

INPrieto <- count(grep("PRIETO S V, 2018, J CLEAN PROD",local_citations30[["CR"]]))
sum(INPrieto$freq)

INMoreno <- count(grep("MORENO M, 2016, SUSTAINABILITY",local_citations30[["CR"]]))
sum(INMoreno$freq)

INMerli <- count(grep("MERLI R, 2018, J CLEAN PROD",local_citations30[["CR"]]))
sum(INMerli$freq)

INElia <- count(grep("ELIA V, 2017, J CLEAN PROD",local_citations30[["CR"]]))
sum(INElia$freq)

# plot historic citation network ------------------------------------------------------------
histPlot(hist_result, n = 30, size = 10, labelsize = 5)

# to reproduce the style of the figure used in the paper a downgrade to bibliometrix version 2.1.0 is required

# # downgrade
# library(devtools)
# install_version("bibliometrix", version = "2.1.0", repos = "http://cran.us.r-project.org")
# library("bibliometrix")
# 
# # plot figure in the format of the paper
# histPlot(hist_result,
#          size.cex =  TRUE,
#          arrowsize = 1,
#          edgesize = 1,
#          size = 20,
#          color = TRUE,
#          labelsize =  40,
#          n = 30)
